pyproject.toml
[tool.poetry]
name = "24-1-drf"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "24"}]

[tool.poetry.dependencies]
python = "^3.11"
django = "4.2.7"
psycopg2 = "^2.9.9"
pillow = "^10.2.0"
djangorestframework = "^3.14.0"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
